test: Obter dados do SIGAA e adequar ao BD
